[new: feature] add metrics plugin
