+++ shenyu-common/src/main/java/org/apache/shenyu/common/config/ShenyuConfig.java
@@ -58,7 +58,27 @@ public class ShenyuConfig {
     private WebsocketConfig websocket = new WebsocketConfig();
 
     private SharedPool sharedPool = new SharedPool();
-
+    
+    private MetricsConfig metrics = new MetricsConfig();
+    
+    /**
+     * Gets metrics.
+     *
+     * @return the metrics
+     */
+    public MetricsConfig getMetrics() {
+        return metrics;
+    }
+    
+    /**
+     * Sets metrics.
+     *
+     * @param metrics the metrics
+     */
+    public void setMetrics(final MetricsConfig metrics) {
+        this.metrics = metrics;
+    }
+    
     /**
      * Gets the shared thread pool config.
      *
@@ -67,7 +87,7 @@ public class ShenyuConfig {
     public SharedPool getSharedPool() {
         return sharedPool;
     }
-
+    
     /**
      * Sets the shared thread pool config.
      *
@@ -76,7 +96,7 @@ public class ShenyuConfig {
     public void setSharedPool(final SharedPool sharedPool) {
         this.sharedPool = sharedPool;
     }
-
+    
     /**
      * Gets the local config.
      *
@@ -103,7 +123,7 @@ public class ShenyuConfig {
     public RibbonConfig getRibbon() {
         return ribbon;
     }
-
+    
     /**
      * Sets ribbon.
      *
@@ -112,7 +132,7 @@ public class ShenyuConfig {
     public void setRibbon(final RibbonConfig ribbon) {
         this.ribbon = ribbon;
     }
-
+    
     /**
      * Gets instance.
      *
@@ -220,7 +240,7 @@ public class ShenyuConfig {
     public void setExclude(final ExcludePath exclude) {
         this.exclude = exclude;
     }
-
+    
     /**
      * Gets fallback.
      *
@@ -229,7 +249,7 @@ public class ShenyuConfig {
     public FallbackPath getFallback() {
         return fallback;
     }
-
+    
     /**
      * Sets fallback.
      *
@@ -238,7 +258,7 @@ public class ShenyuConfig {
     public void setFallback(final FallbackPath fallback) {
         this.fallback = fallback;
     }
-
+    
     /**
      * Gets upstream check.
      *
@@ -265,7 +285,7 @@ public class ShenyuConfig {
     public CrossFilterConfig getCross() {
         return cross;
     }
-
+    
     /**
      * Gets the websocket config.
      *
@@ -274,7 +294,7 @@ public class ShenyuConfig {
     public WebsocketConfig getWebsocket() {
         return websocket;
     }
-
+    
     /**
      * Sets the websocket config.
      *
@@ -512,7 +532,7 @@ public class ShenyuConfig {
             return paths;
         }
     }
-
+    
     /**
      * The type fallback path.
      */
@@ -521,7 +541,7 @@ public class ShenyuConfig {
         private Boolean enabled = false;
 
         private List<String> paths = new ArrayList<>();
-
+    
         /**
          * Gets enabled.
          *
@@ -530,7 +550,7 @@ public class ShenyuConfig {
         public Boolean getEnabled() {
             return enabled;
         }
-
+    
         /**
          * Sets enabled.
          *
@@ -539,7 +559,7 @@ public class ShenyuConfig {
         public void setEnabled(final Boolean enabled) {
             this.enabled = enabled;
         }
-
+    
         /**
          * Sets paths.
          *
@@ -548,7 +568,7 @@ public class ShenyuConfig {
         public void setPaths(final List<String> paths) {
             this.paths = paths;
         }
-
+    
         /**
          * get paths.
          *
@@ -974,7 +994,7 @@ public class ShenyuConfig {
         private String serverLists;
     
         private Properties props = new Properties();
-        
+    
         /**
          * Instantiates a new Instance config.
          */
@@ -1067,7 +1087,7 @@ public class ShenyuConfig {
             this.props = props;
         }
     }
-
+    
     /**
      * The Ribbon Config.
      */
@@ -1077,13 +1097,13 @@ public class ShenyuConfig {
          * see {@code com.netflix.client.config.CommonClientConfigKey#ServerListRefreshInterval}.
          */
         private Integer serverListRefreshInterval = 10000;
-
+    
         /**
          * Instantiates a new RibbonConfig.
          */
         public RibbonConfig() {
         }
-
+    
         /**
          * Instantiates a new RibbonConfig.
          *
@@ -1092,7 +1112,7 @@ public class ShenyuConfig {
         public RibbonConfig(final Integer serverListRefreshInterval) {
             this.serverListRefreshInterval = serverListRefreshInterval;
         }
-
+    
         /**
          * Gets serverListRefreshInterval.
          *
@@ -1101,7 +1121,7 @@ public class ShenyuConfig {
         public Integer getServerListRefreshInterval() {
             return serverListRefreshInterval;
         }
-
+    
         /**
          * setServerListRefreshInterval.
          *
@@ -1120,14 +1140,22 @@ public class ShenyuConfig {
         private Boolean enabled = true;
         
         private String sha512Key;
-        
+    
+        /**
+         * Instantiates a new Local.
+         */
         public Local() {
         }
-        
+    
+        /**
+         * Instantiates a new Local.
+         *
+         * @param sha512Key the sha 512 key
+         */
         public Local(final String sha512Key) {
             this.sha512Key = sha512Key;
         }
-        
+    
         /**
          * Gets enabled.
          *
@@ -1136,7 +1164,7 @@ public class ShenyuConfig {
         public Boolean getEnabled() {
             return enabled;
         }
-        
+    
         /**
          * Sets enabled.
          *
@@ -1145,7 +1173,7 @@ public class ShenyuConfig {
         public void setEnabled(final Boolean enabled) {
             this.enabled = enabled;
         }
-        
+    
         /**
          * Get Sha512Key.
          *
@@ -1154,7 +1182,7 @@ public class ShenyuConfig {
         public String getSha512Key() {
             return sha512Key;
         }
-        
+    
         /**
          * Set Sha512Key.
          *
@@ -1164,7 +1192,7 @@ public class ShenyuConfig {
             this.sha512Key = sha512Key;
         }
     }
-
+    
     /**
      * the websocket config.
      */
@@ -1174,24 +1202,26 @@ public class ShenyuConfig {
          * max frame pay load size mb.
          */
         private Integer maxFramePayloadSize = 10;
-
+    
         /**
          * Get max frame payload size.
+         *
          * @return the max frame payload szie
          */
         public Integer getMaxFramePayloadSize() {
             return maxFramePayloadSize;
         }
-
+    
         /**
          * Set max frame payload size.
+         *
          * @param maxFramePayloadSize the max frame paylod size
          */
         public void setMaxFramePayloadSize(final Integer maxFramePayloadSize) {
             this.maxFramePayloadSize = maxFramePayloadSize;
         }
     }
-
+    
     /**
      * The type Shared Thread Pool.
      */
@@ -1235,7 +1265,7 @@ public class ShenyuConfig {
          * controls memory directly by calculating the memory size used by the blocking queue.
          */
         private Long maxWorkQueueMemory = MemoryLimitCalculator.defaultLimit();
-
+    
         /**
          * Whether to enable shared thread pool.
          *
@@ -1244,7 +1274,7 @@ public class ShenyuConfig {
         public Boolean getEnable() {
             return enable;
         }
-
+    
         /**
          * Set enable.
          *
@@ -1253,7 +1283,7 @@ public class ShenyuConfig {
         public void setEnable(final Boolean enable) {
             this.enable = enable;
         }
-
+    
         /**
          * Get shared thread pool name prefix.
          *
@@ -1262,7 +1292,7 @@ public class ShenyuConfig {
         public String getPrefix() {
             return prefix;
         }
-
+    
         /**
          * Set prefix.
          *
@@ -1271,7 +1301,7 @@ public class ShenyuConfig {
         public void setPrefix(final String prefix) {
             this.prefix = prefix;
         }
-
+    
         /**
          * Get shared thread pool core size.
          *
@@ -1280,7 +1310,7 @@ public class ShenyuConfig {
         public Integer getCorePoolSize() {
             return corePoolSize;
         }
-
+    
         /**
          * Set core pool size.
          *
@@ -1289,7 +1319,7 @@ public class ShenyuConfig {
         public void setCorePoolSize(final Integer corePoolSize) {
             this.corePoolSize = corePoolSize;
         }
-
+    
         /**
          * Get shared thread pool maximum size.
          *
@@ -1298,7 +1328,7 @@ public class ShenyuConfig {
         public Integer getMaximumPoolSize() {
             return maximumPoolSize;
         }
-
+    
         /**
          * Set max pool size.
          *
@@ -1307,7 +1337,7 @@ public class ShenyuConfig {
         public void setMaximumPoolSize(final Integer maximumPoolSize) {
             this.maximumPoolSize = maximumPoolSize;
         }
-
+    
         /**
          * Get shared thread pool keep alive time.
          *
@@ -1316,7 +1346,7 @@ public class ShenyuConfig {
         public Long getKeepAliveTime() {
             return keepAliveTime;
         }
-
+    
         /**
          * Set keep alive time.
          *
@@ -1325,7 +1355,7 @@ public class ShenyuConfig {
         public void setKeepAliveTime(final Long keepAliveTime) {
             this.keepAliveTime = keepAliveTime;
         }
-
+    
         /**
          * Get shared thread pool max work queue memory.
          *
@@ -1334,7 +1364,7 @@ public class ShenyuConfig {
         public Long getMaxWorkQueueMemory() {
             return maxWorkQueueMemory;
         }
-
+    
         /**
          * Set max work queue memory.
          *
@@ -1342,6 +1372,138 @@ public class ShenyuConfig {
          */
         public void setMaxWorkQueueMemory(final Long maxWorkQueueMemory) {
             this.maxWorkQueueMemory = maxWorkQueueMemory;
+        }
+    }
+    
+    /**
+     * The type Metrics config.
+     */
+    public static class MetricsConfig {
+    
+        private boolean enabled;
+        
+        private String name;
+        
+        private String host;
+        
+        private Integer port;
+        
+        private String jmxConfig;
+        
+        private Properties props;
+    
+        /**
+         * Instantiates a new Metrics config.
+         */
+        public MetricsConfig() {
+        }
+    
+        /**
+         * Gets enabled.
+         *
+         * @return the enabled
+         */
+        public boolean getEnabled() {
+            return enabled;
+        }
+    
+        /**
+         * Sets enabled.
+         *
+         * @param enabled the enabled
+         */
+        public void setEnabled(final boolean enabled) {
+            this.enabled = enabled;
+        }
+    
+        /**
+         * Gets metrics name.
+         *
+         * @return the metrics name
+         */
+        public String getName() {
+            return name;
+        }
+    
+        /**
+         * Sets metrics name.
+         *
+         * @param name the metrics name
+         */
+        public void setName(final String name) {
+            this.name = name;
+        }
+    
+        /**
+         * Gets host.
+         *
+         * @return the host
+         */
+        public String getHost() {
+            return host;
+        }
+    
+        /**
+         * Sets host.
+         *
+         * @param host the host
+         */
+        public void setHost(final String host) {
+            this.host = host;
+        }
+    
+        /**
+         * Gets port.
+         *
+         * @return the port
+         */
+        public Integer getPort() {
+            return port;
+        }
+    
+        /**
+         * Sets port.
+         *
+         * @param port the port
+         */
+        public void setPort(final Integer port) {
+            this.port = port;
+        }
+    
+        /**
+         * Gets jmx config.
+         *
+         * @return the jmx config
+         */
+        public String getJmxConfig() {
+            return jmxConfig;
+        }
+    
+        /**
+         * Sets jmx config.
+         *
+         * @param jmxConfig the jmx config
+         */
+        public void setJmxConfig(final String jmxConfig) {
+            this.jmxConfig = jmxConfig;
+        }
+    
+        /**
+         * Gets props.
+         *
+         * @return the props
+         */
+        public Properties getProps() {
+            return props;
+        }
+    
+        /**
+         * Sets props.
+         *
+         * @param props the props
+         */
+        public void setProps(final Properties props) {
+            this.props = props;
         }
     }
 }
